refactor: mock 파일 import하여 사용하도록 변경
chat.service의 경우, 사용자 mock 데이터가 변경되면 다시 수정 필요
