+++ backend/was/src/tarot/tarot.service.spec.ts
@@ -1,38 +1,22 @@
 import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
+import {
+  tarotCardMock,
+  tarotResultId,
+  tarotResultMessage,
+  tarotResultMock,
+} from 'src/mocks/tarot';
 import { Repository } from 'typeorm';
 import { v4 as uuidv4 } from 'uuid';
-import { CreateTarotResultDto } from './dto/create-tarot-result.dto';
-import { TarotCardResponseDto } from './dto/tarot-card-response.dto';
-import { TarotResultResponseDto } from './dto/tarot-result-response.dto';
-import { TarotCard } from './entities/tarot-card.entity';
-import { TarotResult } from './entities/tarot-result.entity';
+import { CreateTarotResultDto, TarotCardDto, TarotResultDto } from './dto';
+import { TarotCard, TarotResult } from './entities';
 import { TarotService } from './tarot.service';
 
 describe('TarotService', () => {
   let service: TarotService;
   let tarotCardRepository: Repository<TarotCard>;
   let tarotResultRepository: Repository<TarotResult>;
-
-  /**
-   * mock data
-   */
-  const tarotResultId = uuidv4();
-
-  const cardUrlMock =
-    'https://kr.object.ncloudstorage.com/magicconch/basic/0.jpg';
-
-  const tarotResultMessageMock = 'tarot result message';
-
-  const tarotCardMock = new TarotCard();
-  tarotCardMock.cardNo = 0;
-  tarotCardMock.ext = '.jpg';
-
-  const tarotResultMock = new TarotResult();
-  tarotResultMock.id = tarotResultId;
-  tarotResultMock.cardUrl = cardUrlMock;
-  tarotResultMock.message = tarotResultMessageMock;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -67,7 +51,7 @@ describe('TarotService', () => {
       const createTarotResultDto: CreateTarotResultDto =
         CreateTarotResultDto.fromResult(
           tarotCardMock.cardNo,
-          tarotResultMessageMock,
+          tarotResultMessage,
         );
 
       const saveMock = jest
@@ -87,8 +71,7 @@ describe('TarotService', () => {
 
   describe('findTarotCardById', () => {
     it('should find specific tarot card', async () => {
-      const tarotCardResponseDto: TarotCardResponseDto =
-        TarotCardResponseDto.fromEntity(tarotCardMock);
+      const tarotCardDto: TarotCardDto = TarotCardDto.fromEntity(tarotCardMock);
 
       const findOneByMock = jest
         .spyOn(tarotCardRepository, 'findOneBy')
@@ -96,7 +79,7 @@ describe('TarotService', () => {
 
       const result = await service.findTarotCardById(tarotCardMock.cardNo);
 
-      expect(result).toEqual(tarotCardResponseDto);
+      expect(result).toEqual(tarotCardDto);
 
       expect(findOneByMock).toHaveBeenCalledWith({
         cardNo: tarotCardMock.cardNo,
@@ -124,8 +107,7 @@ describe('TarotService', () => {
 
   describe('findTarotResultById', () => {
     it('should find specific tarot result', async () => {
-      const tarotResultResponseDto =
-        TarotResultResponseDto.fromEntity(tarotResultMock);
+      const tarotResultDto = TarotResultDto.fromEntity(tarotResultMock);
 
       const findOneByMock = jest
         .spyOn(tarotResultRepository, 'findOneBy')
@@ -133,7 +115,7 @@ describe('TarotService', () => {
 
       const result = await service.findTarotResultById(tarotResultId);
 
-      expect(result).toEqual(tarotResultResponseDto);
+      expect(result).toEqual(tarotResultDto);
 
       expect(findOneByMock).toHaveBeenCalledWith({ id: tarotResultId });
     });
@@ -143,13 +125,13 @@ describe('TarotService', () => {
         .spyOn(tarotResultRepository, 'findOneBy')
         .mockResolvedValueOnce(null);
 
-      const nonResultId = uuidv4();
+      const wrongResultId = uuidv4();
 
-      await expect(service.findTarotResultById(nonResultId)).rejects.toThrow(
+      await expect(service.findTarotResultById(wrongResultId)).rejects.toThrow(
         NotFoundException,
       );
 
-      expect(findOneByMock).toHaveBeenCalledWith({ id: nonResultId });
+      expect(findOneByMock).toHaveBeenCalledWith({ id: wrongResultId });
     });
   });
 });
